requirements.txt
Flask==2.3.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.84.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
gspread==5.12.4
oauth2client==4.1.3
pytz==2023.3  # For timezone handling
